requirements.txt
tornado==5.1
wtforms
peewee
pymysql
aiomysql
PyJWT
aiofiles
pillow
peewee-async
wtforms_tornado
redis
bcrypt
wtforms_json